Lot corrections/évolutions
db/drop_db.sh
    Test erreur dbca

db/drop_pdb.sh
    Ajout tests

Backup scindé pour mieux voir la commande qui échoue
    db/image_copy_backup.sh
    db/rman/image_copy.rman
    db/rman/backup_archive_log.rman
    db/rman/set_config.rman (active la compression)

db/remove_all_files_for_db.sh
    Ne fait le kill que si c'est nécessaire.

db/stby/create_dataguard.sh
    Active la compression pendant le duplicate.

db/stby/drop_stby_cfg.sh
    Revue du code.

san/check_target.sh
    Message plus parlant.

+++ db/drop_db.sh
@@ -116,6 +116,16 @@ add_dynamic_cmd_param "    -silent"
 exec_dynamic_cmd dbca
 LN
 
+if ps -ef|grep -q [p]mon_$db
+then
+	error "dbca failed."
+	LN
+	info "L'affichage de dbca et son code retour peuvent laissé croire"
+	info "que tout c'est bien passé sauf que parfois dbca redémarre (cf log)"
+	LN
+	exit 1
+fi
+
 typeset -r rm_1="rm -rf $ORACLE_BASE/cfgtoollogs/dbca/${db}*"
 typeset -r rm_2="rm -rf $ORACLE_BASE/diag/rdbms/$(to_lower $db)"
 typeset -r rm_3="rm -rf $ORACLE_BASE/admin/$db"
